test: refactor and move tests from `src` to `__tests__` directory
