+++ .lintstagedrc.json
@@ -1,4 +1,4 @@
 {
-  "*.{js,mjs,ts}": "eslint --fix",
-  "*.{css,json,md,yml}": "prettier --write"
+  "*.{cjs,cts,js,jsx,mjs,mts,ts,tsx}": "eslint --fix",
+  "*.{css,html,json,md,yaml,yml}": "prettier --write"
 }
