+++ __tests__/constants/events.test.ts
@@ -1,4 +1,4 @@
-import { events } from './events';
+import { events } from '../../src/constants/events';
 
 it('exports events', () => {
   expect(events).toMatchSnapshot();
