+++ __tests__/element/jsx.test.ts
@@ -1,0 +1,6 @@
+import { createElement } from '../../src/element/create';
+import { jsx } from '../../src/element/jsx';
+
+it('exports jsx', () => {
+  expect(jsx).toBe(createElement);
+});
