+++ __tests__/jsx-runtime/index.test.ts
@@ -1,5 +1,5 @@
-import { createElement } from '..';
-import { Fragment, jsx, jsxs } from '.';
+import { createElement } from '../../src';
+import { Fragment, jsx, jsxs } from '../../src/jsx-runtime';
 
 it.each([jsx, jsxs])('exports %s', (fun) => {
   expect(fun).toBe(createElement);
