+++ __tests__/render/index.test.ts
@@ -1,4 +1,4 @@
-import { render } from './render';
+import { render } from '../../src/render';
 
 it('exports render', () => {
   expect(render).toBeInstanceOf(Function);
