--- src/element/jsx.test.ts
@@ -1,6 +1,0 @@
-import { createElement } from './create';
-import { jsx } from './jsx';
-
-it('exports jsx', () => {
-  expect(jsx).toBe(createElement);
-});
