+++ tsconfig.build.json
@@ -13,6 +13,5 @@
     "outDir": "cjs",
     "jsx": "react-jsx"
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "**/*.test.*"]
+  "include": ["src"]
 }
